+++ plugins/basesamplers/robotconfiguration.h
@@ -1,5 +1,5 @@
 // -*- coding: utf-8 --*
-// Copyright (C) 2006-2011 Rosen Diankov <rosen.diankov@gmail.com>
+// Copyright (C) 2006-2012 Rosen Diankov <rosen.diankov@gmail.com>
 //
 // This file is part of OpenRAVE.
 // OpenRAVE is free software: you can redistribute it and/or modify
@@ -25,16 +25,13 @@ public:
 Samples the robot active configuration space, treats revolute and circular joints appropriately. When creating pass the following parameters::\n\n\
   RobotConfiguration [robot name] [sampler name]\n\n\
 The sampler needs to return values in the range [0,1]. Default sampler is 'mt19937'.\n\
+If the robot active DOFs change, can use the 'TrackActiveSpace' command to automatically update the sampling configuration space. By default this is true.\n\
 ";
+        RegisterCommand("TrackActiveSpace",boost::bind(&RobotConfigurationSampler::TrackActiveSpaceCommand,this,_1,_2),
+                        "Enable/disable the automating updating of the active configuration space.");
         string robotname;
         sinput >> robotname;
         _probot = GetEnv()->GetRobot(robotname);
-        if( !!_probot ) {
-            _updatedofscallback = _probot->RegisterChangeCallback(RobotBase::Prop_RobotActiveDOFs,boost::bind(&RobotConfigurationSampler::_UpdateDOFs,this));
-        }
-        else {
-            RAVELOG_WARN(str(boost::format("failed to find robot '%s'\n")%robotname));
-        }
         string samplername;
         sinput >> samplername;
         if( samplername.size() == 0 ) {
@@ -48,6 +45,9 @@ The sampler needs to return values in the range [0,1]. Default sampler is 'mt199
             for(int i = 0; i < GetDOF(); ++i) {
                 BOOST_ASSERT(vsamplerlower[i] == 0 && vsamplerupper[i] == 1);
             }
+        }
+        if( !!_probot ) {
+            _updatedofscallback = _probot->RegisterChangeCallback(RobotBase::Prop_RobotActiveDOFs,boost::bind(&RobotConfigurationSampler::_UpdateDOFs,this));
         }
     }
 
@@ -98,6 +98,26 @@ The sampler needs to return values in the range [0,1]. Default sampler is 'mt199
     }
 
 protected:
+
+    bool TrackActiveSpaceCommand(ostream& sout, istream& sinput)
+    {
+        bool btrack = false;
+        sinput >> btrack;
+        if( !sinput ) {
+            return false;
+        }
+
+        if( !!_probot && btrack ) {
+            if( !_updatedofscallback ) {
+                _updatedofscallback = _probot->RegisterChangeCallback(RobotBase::Prop_RobotActiveDOFs,boost::bind(&RobotConfigurationSampler::_UpdateDOFs,this));
+            }
+        }
+        else {
+            _updatedofscallback.reset();
+        }
+        return true;
+    }
+
     Vector _SampleQuaternion()
     {
         _tempsamples.resize(4);
@@ -147,8 +167,8 @@ protected:
     SpaceSamplerBasePtr _psampler;
     RobotBasePtr _probot;
     boost::shared_ptr<void> _updatedofscallback;
-    vector<dReal> _lower, _upper, _range, _rangescaled;
-    vector<dReal> _tempsamples;
-    vector<uint8_t> _viscircular;
+    std::vector<dReal> _lower, _upper, _range, _rangescaled;
+    std::vector<dReal> _tempsamples;
+    std::vector<uint8_t> _viscircular;
     int _affinerotaxis, _affinerot3d, _affinequat;
 };