+++ plugins/rplanners/rrt.h
@@ -39,13 +39,14 @@ Uses the Rapidly-Exploring Random Trees Algorithm.\n\
 
     virtual bool _InitPlan(RobotBasePtr pbase, PlannerParametersPtr params)
     {
+        params->Validate();
         _goalindex = -1;
         _startindex = -1;
         EnvironmentMutex::scoped_lock lock(GetEnv()->GetMutex());
         _uniformsampler = RaveCreateSpaceSampler(GetEnv(),"mt19937");
         _robot = pbase;
 
-        RobotBase::RobotStateSaver savestate(_robot);
+        PlannerParameters::StateSaver savestate(params);
         CollisionOptionsStateSaver optionstate(GetEnv()->GetCollisionChecker(),GetEnv()->GetCollisionChecker()->GetCollisionOptions()|CO_ActiveDOFs,false);
 
         if( (int)params->vinitialconfig.size() % params->GetDOF() ) {
@@ -128,10 +129,6 @@ Uses the Rapidly-Exploring Random Trees Algorithm.\n\
         }
     }
 
-    virtual RobotBasePtr GetRobot() const {
-        return _robot;
-    }
-
     bool GetGoalIndexCommand(std::ostream& os, std::istream& is)
     {
         os << _goalindex;
@@ -183,7 +180,7 @@ public:
         }
 
         _fGoalBiasProb = dReal(0.01);
-        RobotBase::RobotStateSaver savestate(_robot);
+        PlannerParameters::StateSaver savestate(_parameters);
         CollisionOptionsStateSaver optionstate(GetEnv()->GetCollisionChecker(),GetEnv()->GetCollisionChecker()->GetCollisionOptions()|CO_ActiveDOFs,false);
 
         _treeBackward.Reset(shared_planner(), _parameters->GetDOF());
@@ -247,7 +244,7 @@ public:
         uint32_t basetime = utils::GetMilliTime();
 
         // the main planning loop
-        RobotBase::RobotStateSaver savestate(_robot);
+        PlannerParameters::StateSaver savestate(_parameters);
         CollisionOptionsStateSaver optionstate(GetEnv()->GetCollisionChecker(),GetEnv()->GetCollisionChecker()->GetCollisionOptions()|CO_ActiveDOFs,false);
 
         SpatialTreeBase* TreeA = &_treeForward;
@@ -573,7 +570,7 @@ public:
         bool bSuccess = false;
 
         // the main planning loop
-        RobotBase::RobotStateSaver savestate(_robot);
+        PlannerParameters::StateSaver savestate(_parameters);
         CollisionOptionsStateSaver optionstate(GetEnv()->GetCollisionChecker(),GetEnv()->GetCollisionChecker()->GetCollisionOptions()|CO_ActiveDOFs,false);
 
         PlannerAction callbackaction = PA_None;
@@ -784,7 +781,7 @@ protected:
         EnvironmentMutex::scoped_lock lock(GetEnv()->GetMutex());
         vector<dReal> vSampleConfig;
 
-        RobotBase::RobotStateSaver saver(_robot);
+        PlannerParameters::StateSaver savestate(_parameters);
         CollisionOptionsStateSaver optionstate(GetEnv()->GetCollisionChecker(),GetEnv()->GetCollisionChecker()->GetCollisionOptions()|CO_ActiveDOFs,false);
 
         int iter = 0;